+++ src/main/java/chess/game/Application.java
@@ -1,6 +1,4 @@
-package chess;
-
-import chess.game.ChessGame;
+package chess.game;
 
 public class Application {
