+++ src/main/java/chess/web/WebApplication.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.web;
 
 import static spark.Spark.get;
 import static spark.Spark.port;
